+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- feature: add `./run/update` for updating an already-provisioned
 - feature: install amdgpu_top
 - feature: install orca screen reader
 - feature: install pied (UI for piper)